dlp: split main into specialized files
