+++ dlp/deid.go
@@ -1,0 +1,89 @@
+package main
+
+import (
+	"context"
+	"fmt"
+	"io"
+	"log"
+
+	dlp "cloud.google.com/go/dlp/apiv2"
+	dlppb "google.golang.org/genproto/googleapis/privacy/dlp/v2"
+)
+
+func mask(w io.Writer, client *dlp.Client, project, s string) {
+	rcr := &dlppb.DeidentifyContentRequest{
+		Parent: "projects/" + project,
+		DeidentifyConfig: &dlppb.DeidentifyConfig{
+			Transformation: &dlppb.DeidentifyConfig_InfoTypeTransformations{
+				InfoTypeTransformations: &dlppb.InfoTypeTransformations{
+					Transformations: []*dlppb.InfoTypeTransformations_InfoTypeTransformation{
+						{
+							InfoTypes: []*dlppb.InfoType{},
+							PrimitiveTransformation: &dlppb.PrimitiveTransformation{
+								Transformation: &dlppb.PrimitiveTransformation_CharacterMaskConfig{
+									CharacterMaskConfig: &dlppb.CharacterMaskConfig{
+										MaskingCharacter: "*",
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+		Item: &dlppb.ContentItem{
+			DataItem: &dlppb.ContentItem_Value{
+				Value: s,
+			},
+		},
+	}
+	r, err := client.DeidentifyContent(context.Background(), rcr)
+	if err != nil {
+		log.Fatal(err)
+	}
+	fmt.Fprintln(w, r.GetItem().GetValue())
+}
+
+func deidentifyFPE(w io.Writer, client *dlp.Client, project, s, wrappedKey, cryptoKeyName string) {
+	rcr := &dlppb.DeidentifyContentRequest{
+		Parent: "projects/" + project,
+		DeidentifyConfig: &dlppb.DeidentifyConfig{
+			Transformation: &dlppb.DeidentifyConfig_InfoTypeTransformations{
+				InfoTypeTransformations: &dlppb.InfoTypeTransformations{
+					Transformations: []*dlppb.InfoTypeTransformations_InfoTypeTransformation{
+						{
+							InfoTypes: []*dlppb.InfoType{},
+							PrimitiveTransformation: &dlppb.PrimitiveTransformation{
+								Transformation: &dlppb.PrimitiveTransformation_CryptoReplaceFfxFpeConfig{
+									CryptoReplaceFfxFpeConfig: &dlppb.CryptoReplaceFfxFpeConfig{
+										CryptoKey: &dlppb.CryptoKey{
+											Source: &dlppb.CryptoKey_KmsWrapped{
+												KmsWrapped: &dlppb.KmsWrappedCryptoKey{
+													WrappedKey:    []byte(wrappedKey),
+													CryptoKeyName: cryptoKeyName,
+												},
+											},
+										},
+										Alphabet: &dlppb.CryptoReplaceFfxFpeConfig_CommonAlphabet{
+											CommonAlphabet: dlppb.CryptoReplaceFfxFpeConfig_ALPHA_NUMERIC,
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+		Item: &dlppb.ContentItem{
+			DataItem: &dlppb.ContentItem_Value{
+				Value: s,
+			},
+		},
+	}
+	r, err := client.DeidentifyContent(context.Background(), rcr)
+	if err != nil {
+		log.Fatal(err)
+	}
+	fmt.Fprintln(w, r.GetItem().GetValue())
+}
